requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
cmake==3.26.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
filelock==3.11.0
fonttools==4.39.2
idna==2.10
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.3.0
kiwisolver==1.4.4
lit==16.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
pyzmq==25.0.2
requests==2.25.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
secml==0.15.3
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.1
torchattacks==3.4.0
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.15.0
